requirements.txt
# Basic libraries
numpy

# Object detection
pyyaml
tflite-support
picamera2

# Mapping
scipy
matplotlib
